extracting diff logic to a util file, so it can be reused later
